+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The Following rules must be followed at all times
 - Added Email Verification feature
 - Added Session management feature
 - Added Forgot and Reset Password feature
+- Added Loading States to Session management page
 
 ### Changed
 - Improvised the UI for all Pages